# Server/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Online Food Ordering API
  version: 1.0.0
  description: RESTful API for food ordering system (buyers, sellers, couriers, and admin).
servers:
  - url: http://localhost:8080
...
paths:
  /auth/register:
    post:
      tags:
        - auth
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - full_name
                - phone
                - password
                - role
                - address
              properties:
                full_name:
                  type: string
                phone:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                role:
                  type: string
                  enum:
                    - buyer
                    - seller
                    - courier
                address:
                  type: string
                profileImageBase64:
                  type: string
                bank_info:
                  type: object
                  required:
                    - bankName
                    - accountNumber
                  properties:
                    bank_name:
                      type: string
                    account_number:
                      type: string
      responses:
        '200':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user_id:
                    type: string
                  token:
                    type: string
        '400':
          $ref: '#/components/responses/invalid_input'
        '409':
          description: Phone number already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              examples:
                Phone number already exists:
                  value: |-
                    {
                    "error":"Phone number already exists"
                    }
  /auth/login:
    post:
      tags:
        - auth
      summary: Login a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phone
                - password
              properties:
                phone:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: User logged in successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  token:
                    type: string
                  user:
                    $ref: "#/components/schemas/user"
        '401':
          $ref: '#/components/responses/unauthorized'
  /auth/profile:
    get:
      tags:
        - auth
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/user"
        '401':
          $ref: '#/components/responses/unauthorized'
    put:
      tags:
        - auth
      summary: Update current user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/user"
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/message"
        '400':
          $ref: '#/components/responses/invalid_input'
        '401':
          $ref: '#/components/responses/unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    message:
      type: object
      required: ['message']
      properties:
        message:
          type: string
    user:
      type: object
      required: ["id", "full_name", "phone", "email", "role", "address"]
      properties:
        id:
          type: string
        full_name:
          type: string
        phone:
          type: string
        email:
          type: string
        role:
          type: string
        address:
          type: string
        profileImageBase64:
          type: string
        bank_info:
          type: object
          required: ["bank_name", "account_number"]
          properties:
            bank_name:
              type: string
            account_number:
              type: string
...
  responses:
    unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
          examples:
            Unauthorized:
              value: |-
                {
                "error":"Unauthorized request"
                }
    invalid_input:
      description: Invalid input
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
          examples:
            Invalid Field:
              value: |-
                {
                "error":"Invalid `field name`"
                }